// components/rent_page/add_info_block/enums.ts
export enum Concierge {
  HEADER = "Консьерж сервис",
  DESC = "Сэкономим ваше время и нервы. Подберём квартиры по вашему запросу, договоримся о просмотрах",
  IMAGE = "/images/man-sits.webp",
  CARD_TYPE = "concierge",
}

export enum Owner {
  HEADER = "Как подготовиться к встрече с собственником?",
  DESC = "Разработали пакет документов с чек-листами по осмотру объекта и вопросами собственнику",
  IMAGE = "/images/man-stands.webp",
  CARD_TYPE = "owner",
}